backend/requirements.txt
fastapi
uvicorn[standard]
onnxruntime
nltk
numpy
pydantic
huggingface_hub
python-dotenv
